Twerks to base locator.
We start with a raster with large cells and refine. While this should make
little difference to the result, it constrains memory consumption by
keeping the number of cells in the raster in any iteration more bounded.

We also allows the alignment of the grid to vary, which may improve
accuracy, but did not change anything in the (one) test I ran.